basic autoconf/automake/libtool support from Albert Chin-A-Young.
I made three changes:

updated for Makefile.am with the new source filename prefix

removed spurious X11 discovery

wrote an autogen.sh and re-arranged/renamed some files.
typically, one doesn't check the generated code for the configure system
into cvs; one just puts in the source and a script called 'autogen.sh'
that takes care of building things to the point where the traditional

./configure
make
make install

will work. To do otherwise generates a lot of unnecessary traffic and
spurious messages as cvs tries to merge the changes between all the
slight-different versions of the object code.

Also strange here is that I've sequestered most of said source in the
'makefiles' subdirectory. In a *nix-centric project, they'd all be on
the top level, but Gerard and I wished to avoid the clutter from the
point of view of other platforms.

./autogen.sh will copy (link, actually) the missing files into their
conventional places, so things will work as normal, and the .tar.gz
distributions should look like unix projects. For .zip dists, we just
don't run autogen, and things stay quiet.

This is pretty simple support so far. It doesn't, for example, figure
out the correct endianness or type sizes. autogen.sh needs more error
checking. that sort of thing.