This should resolve bug #117103
(https://sourceforge.net/bugs/?func=detailbug&group_id=5635&bug_id=117103)

A complicated one; several things were wonky.

The unary -eq error was a missing variable declaration. That's fixed.

The warnings about redefining PACKAGE and VERSION were because lcms
includes them (incorrectly) is lcms.h, which is also their AC_CONFIG
file. I've sent a note to the developers about that.

With that fixed, autodetection *still* fails, because the linux version
of lcms installs the header in lcms/lcms.h not at the top level like
(apparently) the win32 version. Unfortunately, I don't know a clean way
to check for the same header in two locations. This patch is an ugly
hack: if #include <lcms.h> doesn't work, we look explicitly for
/usr/include/lcms/lcms.h and then /usr/local/include/lcms/lcms.h and
add the appropriate -I line to CPPFLAGS if found. That's responsible
for the cross-compilation warnings from autoconf.

The real sol'n is probably to do:

#ifdef MNG_LCMS_SUBDIR
#include <lcms/lcms.h>
#else
#include <lcms.h>
#endif

in all the source files. I thought I'd see if we could convince the
windows side to switch to lcms/lcms.h first. Unless Albert has better
idea?

As a workaround for all this passing --with-lcms=/usr/local (or whatever)
still works fine.

Finally, we now also define MNG_FULL_CMS along with MNG_INCLUDE_LCMS,
which we weren't doing but the code expects. Gerard, from the Linux
point of view, I think it makes more sense to put these define
dependencies in configure.in as we discussed before.